requirements.txt
# Client dependencies (pfSense)
# websockets removed (HTTP mode)
requests>=2.31.0
psutil>=5.9.5
cryptography>=41.0.3
pyyaml>=6.0.1

# Server dependencies (HQ)
fastapi>=0.103.0
uvicorn>=0.23.0
websockets>=11.0.3
openai>=1.3.0
asyncio-mqtt>=0.13.0
sqlalchemy>=2.0.20
alembic>=1.12.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
aiosqlite>=0.19.0